Cleaning up unneeded files and code
