--- backend/zip/dashboard_export_20231111T214531/charts/Popular_Genres_Across_Platforms_74.yaml
@@ -1,39 +1,0 @@
-slice_name: Popular Genres Across Platforms
-description: null
-certified_by: null
-certification_details: null
-viz_type: heatmap
-params:
-  adhoc_filters: []
-  all_columns_x: platform
-  all_columns_y: genre
-  bottom_margin: auto
-  canvas_image_rendering: pixelated
-  datasource: 64__table
-  granularity_sqla: year
-  left_margin: auto
-  linear_color_scheme: blue_white_yellow
-  metric: count
-  normalize_across: heatmap
-  queryFields:
-    metric: metrics
-  row_limit: 10000
-  show_legend: true
-  show_perc: true
-  show_values: true
-  sort_x_axis: alpha_asc
-  sort_y_axis: alpha_asc
-  time_range: No filter
-  url_params: {}
-  viz_type: heatmap
-  xscale_interval: null
-  y_axis_bounds:
-  - null
-  - null
-  y_axis_format: SMART_NUMBER
-  yscale_interval: null
-query_context: null
-cache_timeout: null
-uuid: 326fc7e5-b7f1-448e-8a6f-80d0e7ce0b64
-version: 1.0.0
-dataset_uuid: 53d47c0c-c03d-47f0-b9ac-81225f808283
