--- backend/zip/dashboard_export_20231111T214531/charts/Rise__Fall_of_Video_Game_Consoles_80.yaml
@@ -1,120 +1,0 @@
-slice_name: Rise & Fall of Video Game Consoles
-description: null
-certified_by: null
-certification_details: null
-viz_type: area
-params:
-  adhoc_filters: []
-  annotation_layers: []
-  bottom_margin: auto
-  color_scheme: supersetColors
-  comparison_type: values
-  contribution: false
-  datasource: 21__table
-  granularity_sqla: year
-  groupby:
-  - platform
-  label_colors:
-    '0': '#1FA8C9'
-    '1': '#454E7C'
-    '2600': '#666666'
-    3DO: '#B2B2B2'
-    3DS: '#D1C6BC'
-    Action: '#1FA8C9'
-    Adventure: '#454E7C'
-    DC: '#A38F79'
-    DS: '#8FD3E4'
-    Europe: '#5AC189'
-    Fighting: '#5AC189'
-    GB: '#FDE380'
-    GBA: '#ACE1C4'
-    GC: '#5AC189'
-    GEN: '#3CCCCB'
-    GG: '#EFA1AA'
-    Japan: '#FF7F44'
-    Microsoft Game Studios: '#D1C6BC'
-    Misc: '#FF7F44'
-    N64: '#1FA8C9'
-    NES: '#9EE5E5'
-    NG: '#A1A6BD'
-    Nintendo: '#D3B3DA'
-    North America: '#666666'
-    Other: '#E04355'
-    PC: '#EFA1AA'
-    PCFX: '#FDE380'
-    PS: '#A1A6BD'
-    PS2: '#FCC700'
-    PS3: '#3CCCCB'
-    PS4: '#B2B2B2'
-    PSP: '#FEC0A1'
-    PSV: '#FCC700'
-    Platform: '#666666'
-    Puzzle: '#E04355'
-    Racing: '#FCC700'
-    Role-Playing: '#A868B7'
-    SAT: '#A868B7'
-    SCD: '#8FD3E4'
-    SNES: '#454E7C'
-    Shooter: '#3CCCCB'
-    Simulation: '#A38F79'
-    Sports: '#8FD3E4'
-    Strategy: '#A1A6BD'
-    TG16: '#FEC0A1'
-    Take-Two Interactive: '#9EE5E5'
-    WS: '#ACE1C4'
-    Wii: '#A38F79'
-    WiiU: '#E04355'
-    X360: '#A868B7'
-    XB: '#D3B3DA'
-    XOne: '#FF7F44'
-  line_interpolation: linear
-  metrics:
-  - aggregate: SUM
-    column:
-      column_name: global_sales
-      description: null
-      expression: null
-      filterable: true
-      groupby: true
-      id: 887
-      is_dttm: false
-      optionName: _col_Global_Sales
-      python_date_format: null
-      type: DOUBLE PRECISION
-      verbose_name: null
-    expressionType: SIMPLE
-    hasCustomLabel: false
-    isNew: false
-    label: SUM(Global_Sales)
-    optionName: metric_ufl75addr8c_oqqhdumirpn
-    sqlExpression: null
-  order_desc: true
-  queryFields:
-    groupby: groupby
-    metrics: metrics
-  rich_tooltip: true
-  rolling_type: None
-  row_limit: null
-  show_brush: auto
-  show_legend: false
-  slice_id: 659
-  stacked_style: stream
-  time_grain_sqla: null
-  time_range: No filter
-  url_params:
-    preselect_filters: '{"1389": {"platform": ["PS", "PS2", "PS3", "PS4"], "genre":
-      null, "__time_range": "No filter"}}'
-  viz_type: area
-  x_axis_format: smart_date
-  x_axis_label: Year Published
-  x_axis_showminmax: true
-  x_ticks_layout: auto
-  y_axis_bounds:
-  - null
-  - null
-  y_axis_format: SMART_NUMBER
-query_context: null
-cache_timeout: null
-uuid: 83b0e2d0-d38b-d980-ed8e-e1c9846361b6
-version: 1.0.0
-dataset_uuid: 53d47c0c-c03d-47f0-b9ac-81225f808283
