--- backend/zip/dashboard_export_20231111T214531/charts/of_Games_That_Hit_100k_in_Sales_By_Release_Year_81.yaml
@@ -1,76 +1,0 @@
-slice_name: '# of Games That Hit 100k in Sales By Release Year'
-description: null
-certified_by: null
-certification_details: null
-viz_type: treemap_v2
-params:
-  adhoc_filters: []
-  color_scheme: supersetColors
-  datasource: 21__table
-  granularity_sqla: year
-  groupby:
-  - platform
-  label_colors:
-    '2600': '#D1C6BC'
-    3DO: '#A38F79'
-    3DS: '#B2B2B2'
-    Action: '#ACE1C4'
-    Adventure: '#5AC189'
-    COUNT(*): '#1FA8C9'
-    DC: '#666666'
-    DS: '#E04355'
-    Fighting: '#D1C6BC'
-    GB: '#A1A6BD'
-    GBA: '#A868B7'
-    GC: '#D3B3DA'
-    GEN: '#FF7F44'
-    GG: '#8FD3E4'
-    Microsoft Game Studios: '#FCC700'
-    Misc: '#D3B3DA'
-    N64: '#EFA1AA'
-    NES: '#FEC0A1'
-    NG: '#FCC700'
-    Nintendo: '#666666'
-    PC: '#8FD3E4'
-    PCFX: '#A1A6BD'
-    PS: '#FCC700'
-    PS2: '#454E7C'
-    PS3: '#FF7F44'
-    PS4: '#A38F79'
-    PSP: '#3CCCCB'
-    PSV: '#454E7C'
-    Platform: '#FDE380'
-    Puzzle: '#454E7C'
-    Racing: '#9EE5E5'
-    Role-Playing: '#EFA1AA'
-    SAT: '#5AC189'
-    SCD: '#E04355'
-    SNES: '#FDE380'
-    Shooter: '#B2B2B2'
-    Simulation: '#1FA8C9'
-    Sports: '#FEC0A1'
-    Strategy: '#FF7F44'
-    TG16: '#3CCCCB'
-    Take-Two Interactive: '#E04355'
-    WS: '#A868B7'
-    Wii: '#666666'
-    WiiU: '#1FA8C9'
-    X360: '#5AC189'
-    XB: '#ACE1C4'
-    XOne: '#9EE5E5'
-  metric: count
-  number_format: SMART_NUMBER
-  queryFields:
-    groupby: groupby
-    metrics: metrics
-  row_limit: 10
-  slice_id: 661
-  time_range: No filter
-  treemap_ratio: 1.618033988749895
-  url_params: {}
-  viz_type: treemap_v2
-query_context: null
-cache_timeout: null
-uuid: 2b69887b-23e3-b46d-d38c-8ea11856c555
-version: 1.0.0
-dataset_uuid: 53d47c0c-c03d-47f0-b9ac-81225f808283
