--- backend/zip/dashboard_export_20231111T214531/metadata.yaml
@@ -1,3 +1,0 @@
-version: 1.0.0
-type: Dashboard
-timestamp: '2023-11-11T21:45:31.597196+00:00'
