--- backend/zip/dashboard_export_20231111T214620/charts/Most_Dominant_Platforms_53.yaml
@@ -1,63 +1,0 @@
-slice_name: Most Dominant Platforms
-description: null
-certified_by: null
-certification_details: null
-viz_type: pie
-params:
-  adhoc_filters:
-  - clause: WHERE
-    comparator: '25'
-    expressionType: SIMPLE
-    filterOptionName: filter_znxs4o61aod_n5blgxo29r
-    isExtra: false
-    isNew: false
-    operator: <=
-    sqlExpression: null
-    subject: rank
-  color_scheme: supersetColors
-  datasource: 21__table
-  donut: true
-  granularity_sqla: year
-  groupby:
-  - publisher
-  innerRadius: 45
-  label_line: true
-  labels_outside: true
-  metric:
-    aggregate: SUM
-    column:
-      column_name: global_sales
-      description: null
-      expression: null
-      filterable: true
-      groupby: true
-      id: 887
-      is_dttm: false
-      optionName: _col_Global_Sales
-      python_date_format: null
-      type: DOUBLE PRECISION
-      verbose_name: null
-    expressionType: SIMPLE
-    hasCustomLabel: false
-    isNew: false
-    label: SUM(Global_Sales)
-    optionName: metric_k2rqz0zqhkf_49hu4kq9h3u
-    sqlExpression: null
-  number_format: SMART_NUMBER
-  outerRadius: 67
-  label_type: key
-  queryFields:
-    groupby: groupby
-    metric: metrics
-  row_limit: null
-  show_labels: true
-  show_legend: false
-  slice_id: 658
-  time_range: No filter
-  url_params: {}
-  viz_type: pie
-query_context: null
-cache_timeout: null
-uuid: 1810975a-f6d4-07c3-495c-c3b535d01f21
-version: 1.0.0
-dataset_uuid: 53d47c0c-c03d-47f0-b9ac-81225f808283
