--- backend/zip/dashboard_export_20231111T214620/charts/Total_Sales_per_Market_Grouped_by_Genre_88.yaml
@@ -1,169 +1,0 @@
-slice_name: Total Sales per Market (Grouped by Genre)
-description: null
-certified_by: null
-certification_details: null
-viz_type: dist_bar
-params:
-  adhoc_filters: []
-  bar_stacked: true
-  bottom_margin: auto
-  color_scheme: supersetColors
-  columns: []
-  contribution: false
-  datasource: 21__table
-  granularity_sqla: year
-  groupby:
-  - genre
-  label_colors:
-    '0': '#1FA8C9'
-    '1': '#454E7C'
-    '2600': '#666666'
-    3DO: '#B2B2B2'
-    3DS: '#D1C6BC'
-    Action: '#1FA8C9'
-    Adventure: '#454E7C'
-    DC: '#A38F79'
-    DS: '#8FD3E4'
-    Europe: '#5AC189'
-    Fighting: '#5AC189'
-    GB: '#FDE380'
-    GBA: '#ACE1C4'
-    GC: '#5AC189'
-    GEN: '#3CCCCB'
-    GG: '#EFA1AA'
-    Japan: '#FF7F44'
-    Microsoft Game Studios: '#D1C6BC'
-    Misc: '#FF7F44'
-    N64: '#1FA8C9'
-    NES: '#9EE5E5'
-    NG: '#A1A6BD'
-    Nintendo: '#D3B3DA'
-    North America: '#666666'
-    Other: '#E04355'
-    PC: '#EFA1AA'
-    PCFX: '#FDE380'
-    PS: '#A1A6BD'
-    PS2: '#FCC700'
-    PS3: '#3CCCCB'
-    PS4: '#B2B2B2'
-    PSP: '#FEC0A1'
-    PSV: '#FCC700'
-    Platform: '#666666'
-    Puzzle: '#E04355'
-    Racing: '#FCC700'
-    Role-Playing: '#A868B7'
-    SAT: '#A868B7'
-    SCD: '#8FD3E4'
-    SNES: '#454E7C'
-    Shooter: '#3CCCCB'
-    Simulation: '#A38F79'
-    Sports: '#8FD3E4'
-    Strategy: '#A1A6BD'
-    TG16: '#FEC0A1'
-    Take-Two Interactive: '#9EE5E5'
-    WS: '#ACE1C4'
-    Wii: '#A38F79'
-    WiiU: '#E04355'
-    X360: '#A868B7'
-    XB: '#D3B3DA'
-    XOne: '#FF7F44'
-  metrics:
-  - aggregate: SUM
-    column:
-      column_name: na_sales
-      description: null
-      expression: null
-      filterable: true
-      groupby: true
-      id: 883
-      is_dttm: false
-      optionName: _col_NA_Sales
-      python_date_format: null
-      type: DOUBLE PRECISION
-      verbose_name: null
-    expressionType: SIMPLE
-    hasCustomLabel: true
-    isNew: false
-    label: North America
-    optionName: metric_3pl6jwmyd72_p9o4j2xxgyp
-    sqlExpression: null
-  - aggregate: SUM
-    column:
-      column_name: eu_sales
-      description: null
-      expression: null
-      filterable: true
-      groupby: true
-      id: 884
-      is_dttm: false
-      optionName: _col_EU_Sales
-      python_date_format: null
-      type: DOUBLE PRECISION
-      verbose_name: null
-    expressionType: SIMPLE
-    hasCustomLabel: true
-    isNew: false
-    label: Europe
-    optionName: metric_e8rdyfxxjdu_6dgyhf7xcne
-    sqlExpression: null
-  - aggregate: SUM
-    column:
-      column_name: jp_sales
-      description: null
-      expression: null
-      filterable: true
-      groupby: true
-      id: 885
-      is_dttm: false
-      optionName: _col_JP_Sales
-      python_date_format: null
-      type: DOUBLE PRECISION
-      verbose_name: null
-    expressionType: SIMPLE
-    hasCustomLabel: true
-    isNew: false
-    label: Japan
-    optionName: metric_6gesefugzy6_517l3wowdwu
-    sqlExpression: null
-  - aggregate: SUM
-    column:
-      column_name: other_sales
-      description: null
-      expression: null
-      filterable: true
-      groupby: true
-      id: 886
-      is_dttm: false
-      optionName: _col_Other_Sales
-      python_date_format: null
-      type: DOUBLE PRECISION
-      verbose_name: null
-    expressionType: SIMPLE
-    hasCustomLabel: true
-    isNew: false
-    label: Other
-    optionName: metric_cf6kbre28f_2sg5b5pfq5a
-    sqlExpression: null
-  order_bars: false
-  queryFields:
-    columns: groupby
-    groupby: groupby
-    metrics: metrics
-  row_limit: null
-  show_bar_value: false
-  show_controls: true
-  show_legend: true
-  slice_id: 3548
-  time_range: No filter
-  url_params:
-    preselect_filters: '{"1389": {"platform": ["PS", "PS2", "PS3", "PS4"], "genre":
-      null, "__time_range": "No filter"}}'
-  viz_type: dist_bar
-  x_axis_label: Genre
-  x_ticks_layout: flat
-  y_axis_format: SMART_NUMBER
-query_context: null
-cache_timeout: null
-uuid: d8bf948e-46fd-4380-9f9c-a950c34bcc92
-version: 1.0.0
-dataset_uuid: 53d47c0c-c03d-47f0-b9ac-81225f808283
