--- backend/zip/dashboard_export_20231111T214620/charts/Video_Game_Sales_filter_47.yaml
@@ -1,43 +1,0 @@
-slice_name: Video Game Sales filter
-description: null
-certified_by: null
-certification_details: null
-viz_type: filter_box
-params:
-  adhoc_filters: []
-  datasource: 21__table
-  date_filter: true
-  filter_configs:
-  - asc: true
-    clearable: true
-    column: platform
-    key: s3ItH9vhG
-    label: Platform
-    multiple: true
-    searchAllOptions: false
-  - asc: true
-    clearable: true
-    column: genre
-    key: 202hDeMsG
-    label: Genre
-    multiple: true
-    searchAllOptions: false
-  - asc: true
-    clearable: true
-    column: publisher
-    key: 5Os6jsJFK
-    label: Publisher
-    multiple: true
-    searchAllOptions: false
-  granularity_sqla: year
-  queryFields: {}
-  time_range: No filter
-  url_params:
-    preselect_filters: '{"1389": {"platform": ["PS", "PS2", "PS3", "PS4"], "genre":
-      null, "__time_range": "No filter"}}'
-  viz_type: filter_box
-query_context: null
-cache_timeout: null
-uuid: fd9ce7ec-ae08-4f71-93e0-7c26b132b2e6
-version: 1.0.0
-dataset_uuid: 53d47c0c-c03d-47f0-b9ac-81225f808283
