--- backend/zip/dashboard_export_20231111T214620/datasets/examples/video_game_sales.yaml
@@ -1,163 +1,0 @@
-table_name: video_game_sales
-main_dttm_col: null
-description: null
-default_endpoint: null
-offset: 0
-cache_timeout: null
-schema: main
-sql: ''
-params:
-  remote_id: 64
-  database_name: examples
-  import_time: 1606677834
-template_params: null
-filter_select_enabled: true
-fetch_values_predicate: null
-extra: null
-normalize_columns: false
-uuid: 53d47c0c-c03d-47f0-b9ac-81225f808283
-metrics:
-- metric_name: count
-  verbose_name: COUNT(*)
-  metric_type: null
-  expression: COUNT(*)
-  description: null
-  d3format: null
-  currency: null
-  extra: null
-  warning_text: null
-columns:
-- column_name: year
-  verbose_name: null
-  is_dttm: true
-  is_active: null
-  type: BIGINT
-  advanced_data_type: null
-  groupby: true
-  filterable: true
-  expression: null
-  description: null
-  python_date_format: '%Y'
-  extra: null
-- column_name: global_sales
-  verbose_name: null
-  is_dttm: false
-  is_active: null
-  type: FLOAT64
-  advanced_data_type: null
-  groupby: true
-  filterable: true
-  expression: null
-  description: null
-  python_date_format: null
-  extra: null
-- column_name: na_sales
-  verbose_name: null
-  is_dttm: false
-  is_active: null
-  type: FLOAT64
-  advanced_data_type: null
-  groupby: true
-  filterable: true
-  expression: null
-  description: null
-  python_date_format: null
-  extra: null
-- column_name: other_sales
-  verbose_name: null
-  is_dttm: false
-  is_active: null
-  type: FLOAT64
-  advanced_data_type: null
-  groupby: true
-  filterable: true
-  expression: null
-  description: null
-  python_date_format: null
-  extra: null
-- column_name: eu_sales
-  verbose_name: null
-  is_dttm: false
-  is_active: null
-  type: FLOAT64
-  advanced_data_type: null
-  groupby: true
-  filterable: true
-  expression: null
-  description: null
-  python_date_format: null
-  extra: null
-- column_name: jp_sales
-  verbose_name: null
-  is_dttm: false
-  is_active: null
-  type: FLOAT64
-  advanced_data_type: null
-  groupby: true
-  filterable: true
-  expression: null
-  description: null
-  python_date_format: null
-  extra: null
-- column_name: rank
-  verbose_name: null
-  is_dttm: false
-  is_active: null
-  type: BIGINT
-  advanced_data_type: null
-  groupby: true
-  filterable: true
-  expression: null
-  description: null
-  python_date_format: null
-  extra: null
-- column_name: name
-  verbose_name: null
-  is_dttm: false
-  is_active: null
-  type: STRING
-  advanced_data_type: null
-  groupby: true
-  filterable: true
-  expression: null
-  description: null
-  python_date_format: null
-  extra: null
-- column_name: platform
-  verbose_name: null
-  is_dttm: false
-  is_active: null
-  type: STRING
-  advanced_data_type: null
-  groupby: true
-  filterable: true
-  expression: null
-  description: null
-  python_date_format: null
-  extra: null
-- column_name: publisher
-  verbose_name: null
-  is_dttm: false
-  is_active: null
-  type: STRING
-  advanced_data_type: null
-  groupby: true
-  filterable: true
-  expression: null
-  description: null
-  python_date_format: null
-  extra: null
-- column_name: genre
-  verbose_name: null
-  is_dttm: false
-  is_active: null
-  type: STRING
-  advanced_data_type: null
-  groupby: true
-  filterable: true
-  expression: null
-  description: null
-  python_date_format: null
-  extra: null
-version: 1.0.0
-database_uuid: a2dc77af-e654-49bb-b321-40f6b559a1ee
